Implement basic user input and print game field
Stage 2/5: The user is the gamemaster
Objectives:
In this stage, you should write a program that:
Reads 9 symbols from the input and writes an appropriate 3x3 field. Elements of the field can contain only 'X', 'O' and '_' symbols.
Sets the field to a specific format, i.e. field should start and end with ---------, all lines in between should start and end with '|'
symbol and everything in the middle should be separated with a single space.