Implement and check basic coordinates input
Stage 4/5: First move!
Objectives:
The program should work in the following way:
 - Get the 3x3 field from the input as in the previous stages.
 - Output this 3x3 field with cells before the user's move.
 - Then ask the user about his next move.
 - Then the user should input 2 numbers that represent the cell on which user wants to make his X or O. (9 symbols representing
   the field would be on the first line and these 2 numbers would be on the second line of the user input)
 - Analyze user input and show messages in the following situations:
	"This cell is occupied! Choose another one!" - if the cell is not empty;
	"You should enter numbers!" - if the user enters other symbols;
	"Coordinates should be from 1 to 3!" - if the user goes beyond the field.
 - Then output the table including the user's most recent move.
The program should also check user input. If the user input is unsuitable, the program should ask him to enter coordinates again.
So, you need to output a field from the first line of the input and then ask the user to enter a move. Keep asking until the user
enters coordinate that represents an empty cell on the field and after that output the field with that move. You should output the
field only 2 times - before the move and after a legal move.