Break the terminal's black-in-white
Stage 6/6: Formatted output
Keep the functionality from the previous stages, but now make all links in your browser blue. The Colorama library can help you with this task. This library is already installed in the project so you can use it. With Colorama, all you need to do is read the documentation and you should be able to complete this task!